Create separate files for nested classes
