+++ src/main/java/Duke.java
@@ -6,94 +6,6 @@ public class Duke {
     public static Task[] tasks;
     public static Integer counter1 = 0;
 
-    public static class Task {
-        protected String description;
-        protected boolean isDone;
-        protected Integer order;
-
-        public Task(String description) throws DukeException1 {
-            if(description.equals("blah")) {
-                throw new DukeException1();
-            }
-            this.description = description;
-            this.isDone = false;
-            this.order = 0;
-        }
-
-        public String getStatusIcon() {
-            return (isDone ? "X" : " "); // mark done task with X
-        }
-
-        public void markAsDone() {
-            this.isDone = true;
-        }
-
-        @Override
-        public String toString() {
-            return "[" + this.getStatusIcon() + "] " + this.description;
-        }
-    }
-
-    public static class Deadline extends Task {
-
-        protected String by;
-
-        public Deadline(String description, String by) throws DukeException1{
-            super(description);
-            if(description.equals("deadline")) {
-                throw new DukeException1();
-            }
-            this.by = by;
-        }
-
-        @Override
-        public String toString() {
-            return "[D]" + super.toString() + " (by: " + by + ")";
-        }
-    }
-
-    public static class ToDo extends Task {
-
-        public ToDo(String description) throws DukeException1 {
-            super(description);
-            if(description.equals("todo")) {
-                throw new DukeException1();
-            }
-        }
-
-        @Override
-        public String toString() {
-            return "[T]" + super.toString();
-        }
-    }
-
-    public static class Event extends Task {
-
-        protected String at;
-
-        public Event(String description, String at) throws DukeException1{
-            super(description);
-            if(description.equals("event")) {
-                throw new DukeException1();
-            }
-            this.at = at;
-        }
-
-        @Override
-        public String toString() {
-            return "[E]" + super.toString() + " (at: " + at + ")";
-        }
-    }
-
-    public static class DukeException1 extends Exception {
-        DukeException1() {
-            super();
-        }
-        @Override
-        public String toString() {
-            return "OOPS!!! The description of a todo cannot be empty.";
-        }
-    }
 
     public static String niceLine() {
         return "______________________________________________________________\n";
