+++ src/main/java/DukeException1.java
@@ -1,0 +1,11 @@
+import java.lang.Exception;
+public class DukeException1 extends Exception {
+    DukeException1() {
+        super();
+    }
+
+    @Override
+    public String toString() {
+        return "OOPS!!!";
+    }
+}
