+++ text-ui-test/EXPECTED.TXT
@@ -5,22 +5,48 @@ Hello from
 | |_| | |_| |   <  __/
 |____/ \__,_|_|\_\___|
 
+______________________________________________________________
 Hello! I'm Duke 
-What can I do for you? 
+What can I do for you?
+______________________________________________________________
 
+______________________________________________________________
 	Got it. I've added this task: 
- 	 	[T][ ] borrow book
- 	 Now you have 1 tasks in the list.
-	 Here are the tasks in your list:
+		[T][ ] borrow book
+	Now you have 1 tasks in the list.
+
+______________________________________________________________
+
+______________________________________________________________
+	Here are the tasks in your list:
 	1.[T][ ] borrow book
+
+______________________________________________________________
+
+______________________________________________________________
 	Got it. I've added this task: 
- 	 	[D][ ] return book (by: sunday)
- 	 Now you have 2 tasks in the list.
+		[D][ ] return book (by: sunday)
+	Now you have 2 tasks in the list.
+
+______________________________________________________________
+
+______________________________________________________________
 	Got it. I've added this task: 
- 	 	[E][ ] project meeting (at: Mon 2-4pm)
- 	 Now you have 3 tasks in the list.
-	 Here are the tasks in your list:
+		[E][ ] project meeting (at: Mon 2-4pm)
+	Now you have 3 tasks in the list.
+
+______________________________________________________________
+
+______________________________________________________________
+	Here are the tasks in your list:
 	1.[T][ ] borrow book
 	2.[D][ ] return book (by: sunday)
 	3.[E][ ] project meeting (at: Mon 2-4pm)
-Bye. Hope to see you again soon!
+
+______________________________________________________________
+
+______________________________________________________________
+	Bye. Hope to see you again soon!
+
+______________________________________________________________
+
